Chore : The filter is working (supress background noise, exept the cough in the background)

+++ reducer/noise.go
@@ -6,7 +6,7 @@ func LowPassFilter(input []float64, alpha float64) []float64 {
 		return output
 	}
 	for i := 1; i < len(input); i++ {
-		output[i] = alpha*input[i] + (1-alpha)*output[i-1]
+		output[i] = alpha*input[i] + (float64(1)-alpha)*output[i-1]
 	}
 
 	return output
